requirements.txt
einops==0.8.0
h5py==3.11.0
matplotlib==3.7.5
numpy==1.24.4
opencv_python==4.9.0.80
opencv_python_headless==4.10.0.84
openslide_python==1.3.1
pandas==2.0.3
Pillow==11.0.0
pyvips==2.2.3
PyYAML==6.0.2
scikit_learn==1.3.2
scipy==1.14.1
seaborn==0.13.2
timm==0.5.4
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
transformers==4.39.3
